Time: 709 ms (10.12%), Space: 13.6 MB (32.56%) - LeetHub

+++ 0204-count-primes/0204-count-primes.cpp
@@ -2,7 +2,7 @@ class Solution {
     long long sieveOfEratosthenes(long long n) {
     bool prime[n + 1];
     memset(prime, true, sizeof(prime));
-    for (int p = 2; p * p <= n; p++) {
+    for (long long p = 2; p * p <= n; p++) {
         if (prime[p] == true) {
             for (long long i = p * p; i <= n; i += p) {
                 prime[i] = false;
